requirements.txt
alpaca-py==0.8.2
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.24.3
